New main object available in a script: slimer
slimer provides the same API of the object phantom of PhantomJS.
a phantom object is also available to be compatible with existing scripts.
The single difference between with slimer and phantom, is the version property.
phantom.version returns the version to which SlimerJS is compatible, 1.7, although
slimer.version returns the version of SlimerJS.

The API provided by slimer and phantom is not yet implemented.