Update bin/slimerjs.bat
Added -h and --help options, this could probably be handled in slimer itself by dumping some message or displaying some fancy XUL. All this weird escaping strings in batch files is really a pain.

Added subroutine to search in the default locations for Firefox on 32 and 64 bit systems. If the environment variable SLIMERJSLAUNCHER is not set, :findFirefox will see if firefox.exe exists. If firefox.exe can be found in its usual location then SLIMERJSLAUNCHER will be set to that location.

When verifying the environment variable SLIMERJSLAUNCHER, it will now be determined whether or not the path actually exists instead of only checking that SLIMERJSLAUNCHER is not set to an empty string.

The final value of SLIMERJSLAUNCHER will be echoed back to the user.