Added Firefox Console Redirector.exe
Added superfluous looking check on SLIMERJSLAUNCHER. It fixes an issue
where the batch file bails out while checking whether or not the path
exists.

Firefox Console Redirector should behave exactly as we intended the bat
file to behave except the console redirection actually gets the output
back to the console on windows. There aren't any temp files involved, it
just works. I've put the help menu in and as far as I know it can
replace the batch file. All we really need is for a few people to test
it out. I compiled it for "Any CPU" so it should work on 32 and 64 bit
systems.