Fixes a regression: fatal error NS_NO_INTERFACE about nsIFileURl at startup
This error appears only with packaged version.
This was caused during the loading of coffee-script in bootstrap.js:
the id of this module was not considered as an internal package.
The loader tried to retrieve th file path of the module whereas
it is bundled in the zip file omni.ja...